Streamlined play queue loading
